Накидал filereader

+++ BigFileAccessApi/Contracts/Line.cs
@@ -4,6 +4,6 @@
     {
         public long Number { get; set; }
         public long StartOffset { get; set; }
-        public long Length { get; set; }
+        public int Length { get; set; }
     }
 }
